requirements.txt
langchain==0.2.12
langchain-community==0.2.10
langchain-core==0.2.30
langchain-text-splitters==0.2.2

langgraph==0.2.16

chromadb==0.5.5

pypdf==4.3.1

sentence-transformers==3.0.1

python-dotenv==1.0.1

groq==0.9.0
langchain-groq==0.1.9